backend/requirements.txt
django>=5.2.0,<6.0.0
djangorestframework>=3.16.0,<3.17.0
djangorestframework-simplejwt>=5.2.0,<5.4.0
django-cors-headers>=4.7.0,<4.8.0